feat(ex2): restructure test suite
* creates 'tests/' subdirectory for test modules
* creates 'TestUtils' to stash a bunch of common test boilerplate
* moves activity tests to 'tests/ActivitySpec.hs'
* moves workflow tests to 'tests/WorkflowSpec.hs'
* moves test suite entrypoint to 'tests/Spec.hs'
