--- temporal102/ex2/cabal_project
@@ -1,1 +1,0 @@
-packages: .
