+++ temporal102/ex2/tests/Spec.hs
@@ -1,0 +1,12 @@
+module Main where
+
+import qualified ActivitySpec
+import qualified WorkflowSpec
+
+import Test.Hspec
+
+main :: IO ()
+main = hspec do
+  ActivitySpec.spec
+  WorkflowSpec.spec
+
